working version, try to make abstract

+++ src/headers/Entity.h
@@ -10,7 +10,7 @@ public:
 	//Defenition of entity
 	Entity(int x, int y, int width, int height, int movespeed);
 	virtual ~Entity();
-	virtual void Visualise(double angle)=0;	//Visualise entity
+	virtual void Visualise(int angle)=0;	//Visualise entity
 	virtual void Move(RECT box)=0;	//Move entity
 
 	bool CheckCollisions(RECT a, RECT b);	//Check for collision
